Move argument provider to its own file.
