+++ crates/static-analysis-kernel/src/analysis/analyze.rs
@@ -2,7 +2,7 @@ use crate::analysis::file_context::common::get_file_context;
 use crate::analysis::generated_content::is_generated_file;
 use crate::analysis::javascript::execute_rule;
 use crate::analysis::tree_sitter::{get_query_nodes, get_tree};
-use crate::config_file::ArgumentProvider;
+use crate::arguments::ArgumentProvider;
 use crate::model::analysis::{AnalysisOptions, LinesToIgnore};
 use crate::model::common::Language;
 use crate::model::rule::{RuleInternal, RuleResult};
