+++ crates/static-analysis-kernel/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod analysis;
+pub mod arguments;
 pub mod config_file;
 pub mod constants;
 pub mod model;
