+++ crates/static-analysis-server/src/request.rs
@@ -6,7 +6,8 @@ use crate::model::analysis_request::{AnalysisRequest, ServerRule};
 use crate::model::analysis_response::{AnalysisResponse, RuleResponse};
 use crate::model::violation::violation_to_server;
 use kernel::analysis::analyze::analyze;
-use kernel::config_file::{parse_config_file, ArgumentProvider};
+use kernel::arguments::ArgumentProvider;
+use kernel::config_file::parse_config_file;
 use kernel::model::analysis::AnalysisOptions;
 use kernel::model::rule::{Rule, RuleCategory, RuleInternal, RuleSeverity};
 use kernel::path_restrictions::PathRestrictions;
